Octez-l2-libs: octez layer2 libraries

This is a package containing some libraries used by the layer 2 of Octez.

It contains the following libraries:

Octez_smart_rollup: Library for Smart Rollups

Octez_smart_rollup_wasm_benchmark_lib: Smart Rollup WASM benchmark library

Octez_sqlite: SQLite wrappers and helpers

Rollup_node_sqlite_migrations: SQL migrations for the Rollup node store

Tezos_layer2_irmin_context: Irmin context for the smart rollup PVMs

Tezos_layer2_riscv_context: RiscV implementation of the context for Layer2

Tezos_layer2_store: layer2 storage utils

Tezos_scoru_wasm_durable_snapshot: Durable storage reference implementation

Tezos_scoru_wasm_fast: WASM functionality for SCORU Fast Execution

Tezos_scoru_wasm_helpers: Helpers for the smart rollup wasm functionality and debugger

Tezos_scoru_wasm_helpers_functor: Helpers for the smart rollup wasm functionality and debugger (functor)

Tezos_scoru_wasm_helpers_intf: Helpers for the smart rollup wasm functionality and debugger (signatures)

Tezos_scoru_wasm_test_helpers: Helpers for test of the smart rollup wasm functionality

Tezos_wasmer: Wasmer bindings for SCORU WASM

Tezos_webassembly_interpreter_extra: Additional modules from the WebAssembly REPL used in testing