Octez-libs: Octez libraries

This is a package containing some libraries used by the Octez project.

It contains the following libraries:

Aplonk

Bls12_381_hash

Bls12_381_signature

Brassaia

Brassaia_pack

Brassaia_test_helpers

Brassaia_tezos

Distributed_plonk

Epoxy_tx

Kzg: Toolbox for KZG polynomial commitment

Mec

Octez_bls12_381_polynomial: Polynomials over BLS12-381 finite field - Temporary vendored version of Octez

Octez_srs_extraction

Plompiler

Plonk: Plonk zero-knowledge proving system

Polynomial: Polynomials over finite fields

Ppx_brassaia

Ppx_profiler

Qcheck_tezt: Wrapper to register QCheck tests as Tezt tests

Tezos_base: Meta-package and pervasive type definitions for Tezos

Tezos_base_test_helpers: Octez base test helpers

Tezos_clic

Tezos_context: On-disk context abstraction for octez-node

Tezos_context_brassaia: On-disk context abstraction for octez-node (brassaia)

Tezos_crypto: Library with all the cryptographic primitives used by Tezos

Tezos_crypto_dal: DAL cryptographic primitives

Tezos_error_monad: Error monad

Tezos_error_monad_legacy: Error monad (legacy)

Tezos_event_logging: Octez event logging library

Tezos_event_logging_test_helpers: Test helpers for the event logging library

Tezos_expect_helper

Tezos_gossipsub

Tezos_hacl: Thin layer around hacl-star

Tezos_lazy_containers: A collection of lazy containers whose contents is fetched from arbitrary backend on-demand

Tezos_lwt_result_stdlib: error-aware stdlib replacement

Tezos_micheline: Internal AST and parser for the Michelson language

Tezos_p2p_services: Descriptions of RPCs exported by tezos-p2p

Tezos_profiler_backend: Backends for the Octez Profiler

Tezos_rpc: Library of auto-documented RPCs (service and hierarchy descriptions)

Tezos_rpc_http: Library of auto-documented RPCs (http server and client)

Tezos_rpc_http_client: Library of auto-documented RPCs (http client)

Tezos_rpc_http_client_unix: Unix implementation of the RPC client

Tezos_rpc_http_server: Library of auto-documented RPCs (http server)

Tezos_sapling

Tezos_scoru_wasm

Tezos_stdlib: Yet-another local-extension of the OCaml standard library

Tezos_stdlib_unix: Yet-another local-extension of the OCaml standard library (unix-specific fragment)

Tezos_test_helpers

Tezos_tree_encoding: A general-purpose library to encode arbitrary data in Merkle trees

Tezos_version: Version information generated from Git

Tezos_webassembly_interpreter

Tezos_workers: Worker library

Tezt_wrapper

Cohttp

Data_encoding

Opentelemetry_client_cohttp_lwt

Prometheus

Resto